+++ flow/cssom.js
@@ -7,15 +7,3 @@ declare class CSSRuleList {
   length: number;
   item(index: number): CSSStyleRule;
 }
-
-declare interface JssRenderer {
-  createElement(): void;
-  style(): Function;
-  selector(): Function;
-  attach(): Function;
-  detach(): Function;
-  deploy(): Function;
-  insertRule(): Function;
-  deleteRule(rule: CSSStyleRule): boolean;
-  getRules(): Function;
-}