+++ src/PluginsRegistry.js
@@ -1,6 +1,4 @@
 /* @flow */
-import type {Plugin, RuleOptions} from './types'
-
 export default class PluginsRegistry {
   registry: Array<Plugin> = []