+++ src/sheets.js
@@ -1,5 +1,4 @@
 /* @flow */
-
 import SheetsRegistry from './SheetsRegistry'
 
 /**