+++ src/utils/findClassNames.js
@@ -1,13 +1,11 @@
+/* @flow */
 const dotsRegExp = /[.]/g
 const classesRegExp = /[.][^ ,]+/g
 
 /**
  * Get class names from a selector.
- *
- * @param {String} selector
- * @return {String}
  */
-export default function findClassNames(selector) {
+export default function findClassNames(selector: string): string {
   const classes = selector.match(classesRegExp)
 
   if (!classes) return ''