+++ src/utils/findRenderer.js
@@ -1,3 +1,4 @@
+/* @flow */
 import isInBrowser from 'is-in-browser'
 import DomRenderer from '../backends/DomRenderer'
 import VirtualRenderer from '../backends/VirtualRenderer'
@@ -6,12 +7,8 @@ import VirtualRenderer from '../backends/VirtualRenderer'
  * Find proper renderer.
  * Option `virtual` is used to force use of VirtualRenderer even if DOM is
  * detected, used for testing only.
- *
- * @param {Object} options
- * @return {Renderer}
- * @api private
  */
-export default function findRenderer(options = {}) {
+export default function findRenderer(options: StyleSheetOptions | RuleOptions = {}): any {
   if (options.Renderer) return options.Renderer
   const useVirtual = options.virtual || !isInBrowser
   return useVirtual ? VirtualRenderer : DomRenderer