+++ src/utils/toCssValue.js
@@ -1,15 +1,13 @@
-const joinWithSpace = value => value.join(' ')
+/* @flow */
+const joinWithSpace = (value: Array<string | number>): string => value.join(' ')
 
 /**
  * Converts array values to string.
  *
  * `margin: [['5px', '10px']]` > `margin: 5px 10px;`
  * `border: ['1px', '2px']` > `border: 1px, 2px;`
- *
- * @param {Array} value
- * @return {String|Number|Object}
  */
-export default function toCssValue(value) {
+export default function toCssValue(value: any) {
   if (!Array.isArray(value)) return value
 
   // Support space separated values.